+++ src/theme/foundations/colors.ts
@@ -71,7 +71,7 @@ const researchPrimary = "#00838A";
 const researchSecondary = "#006166";
 const researchLibraryLpa = "#005D53";
 const researchLibrarySchomburg = "#A03E31";
-const researchLibrarySchwartzman = brandSecondary;
+const researchLibrarySchwarzman = brandSecondary;
 const whatsOnPrimary = "#242424";
 const whatsOnSecondary = black;
 
@@ -172,9 +172,9 @@ const researchLibraryObj = {
     schomburg: researchLibrarySchomburg,
     "schomburg-05": hexToRGB(researchLibrarySchomburg, 0.05),
     "schomburg-10": hexToRGB(researchLibrarySchomburg, 0.1),
-    schwartzman: researchLibrarySchwartzman,
-    "schwartzman-05": hexToRGB(researchLibrarySchwartzman, 0.05),
-    "schwartzman-10": hexToRGB(researchLibrarySchwartzman, 0.1),
+    schwarzman: researchLibrarySchwarzman,
+    "schwarzman-05": hexToRGB(researchLibrarySchwarzman, 0.05),
+    "schwarzman-10": hexToRGB(researchLibrarySchwarzman, 0.1),
   },
   dark: {
     lpa: hexToRGB(researchLibraryLpa, 0.2),
@@ -183,9 +183,9 @@ const researchLibraryObj = {
     schomburg: hexToRGB(researchLibrarySchomburg, 0.2),
     "schomburg-05": hexToRGB(researchLibrarySchomburg, 0.05),
     "schomburg-10": hexToRGB(researchLibrarySchomburg, 0.1),
-    schwartzman: hexToRGB(researchLibrarySchwartzman, 0.2),
-    "schwartzman-05": hexToRGB(researchLibrarySchwartzman, 0.05),
-    "schwartzman-10": hexToRGB(researchLibrarySchwartzman, 0.1),
+    schwarzman: hexToRGB(researchLibrarySchwarzman, 0.2),
+    "schwarzman-05": hexToRGB(researchLibrarySchwarzman, 0.05),
+    "schwarzman-10": hexToRGB(researchLibrarySchwarzman, 0.1),
   },
 };